+++ nacos-consume-seata/src/main/java/com/seata/controller/TccSeataController.java
@@ -6,6 +6,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
+
 /**
  * @author: laizc
  * @date: created in 2025/10/27
@@ -19,7 +21,7 @@ public class TccSeataController {
     private TccSeataService tccSeataService;
 
     @GetMapping("/seata")
-    public String seata(Integer num) throws Exception {
+    public String seata(BigDecimal num) throws Exception {
         tccSeataService.placeOrder(num);
         return "ok";
     }
